requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
Faker==17.6.0
gunicorn==20.1.0
Markdown==3.4.1
Pillow==9.3.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
